Learned & Implemented Aspect-Oriented Programming (AOP) using SpringBoot
Learned about the difference between Before, After, and Around. Learned the @Aspect and more about the @Component method, how to call a log from another class using AOP, how to implement @Pointcut, what JoinPoints and ProceedingJoinPoints are, etc. The PokemonAspect is able to get any method within PokemonController use it and log it to the log file.